Rename category controller with 'collection' nomenclature
